requirement.txt
accelerate==1.5.2
deepspeed==0.16.7
einops==0.8.1
evaluate==0.4.0
torch==2.0.1
torchvision==0.15.2
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
huggingface_hub==0.29.2
peft==0.15.2
pillow==11.1.0
responses==0.13.3
safetensors==0.5.3
pytorch==2.0.1
tensorboardx==2.6.2.2
bert_score==0.3.9
